macos_support.py
#!/usr/bin/env python3
"""
Linux Spacedesk Clone - macOS Compatibility Module
Provides macOS-specific screen capture and display management.
"""

import sys
import platform
from PIL import ImageGrab
import subprocess

class MacOSDisplayManager:
    """macOS-specific display management"""
    
    def __init__(self):
        if platform.system() != 'Darwin':
            raise OSError("This module is for macOS only")
    
    def get_displays(self):
        """Get list of connected displays"""
        try:
            result = subprocess.run(['system_profiler', 'SPDisplaysDataType'], 
                                  capture_output=True, text=True)
            # Parse output to get display info
            return result.stdout
        except:
            return "Unable to detect displays"
    
    def create_virtual_display(self):
        """Create virtual display on macOS (limited support)"""
        # macOS doesn't easily support virtual displays like X11
        # This would require third-party tools or custom drivers
        print("Virtual display creation on macOS requires additional setup")
        print("Consider using third-party tools like DisplayLink")
        return False
    
    def capture_screen(self):
        """Capture screen on macOS"""
        try:
            # Use PIL ImageGrab (works on macOS)
            screenshot = ImageGrab.grab()
            return screenshot
        except Exception as e:
            print(f"Screen capture failed: {e}")
            return None

# Monkey patch for cross-platform compatibility
if platform.system() == 'Darwin':
    # On macOS, modify the main server to use macOS-specific methods
    import main
    original_init = main.LinuxSpacedeskServer.__init__
    
    def macos_init(self, *args, **kwargs):
        original_init(self, *args, **kwargs)
        self.display_mode = 'macOS'
        self.macos_manager = MacOSDisplayManager()
    
    main.LinuxSpacedeskServer.__init__ = macos_init
    
    original_create = main.LinuxSpacedeskServer.create_virtual_display
    
    def macos_create_virtual_display(self):
        return self.macos_manager.create_virtual_display()
    
    main.LinuxSpacedeskServer.create_virtual_display = macos_create_virtual_display
    
    original_capture = main.LinuxSpacedeskServer.capture_screen
    
    def macos_capture_screen(self):
        return self.macos_manager.capture_screen()
    
    main.LinuxSpacedeskServer.capture_screen = macos_capture_screen