requirements.txt
pillow
opencv-python
numpy
pyautogui
ffmpeg-python
PyQt5
pyinstaller